requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
albumentations==1.3.0
anyio==3.6.2
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
bokeh==2.4.3
cachetools==5.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorcet==3.0.1
comet-ml==3.31.22
comm==0.1.2
configobj==5.0.8
configparser==5.3.0
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dulwich==0.21.2
everett==3.1.0
executing==1.2.0
fastjsonschema==2.16.2
flatten-dict==0.4.2
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
google-auth-oauthlib==0.4.6
google-auth==2.16.0
grpcio==1.51.1
h5py==3.7.0
holoviews==1.15.4
hvplot==0.8.2
idna==3.4
imageio==2.24.0
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipykernel==6.21.1
ipython-genutils==0.2.0
ipython==8.9.0
isoduration==20.11.0
jason==0.1.7
jedi==0.18.2
jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-client==8.0.2
jupyter-core==5.2.0
jupyter-events==0.6.3
jupyter-server-terminals==0.4.4
jupyter-server==2.2.1
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
kornia==0.6.11
lightning-utilities==0.4.2
markdown==3.4.1
markupsafe==2.1.2
matplotlib-inline==0.1.6
matplotlib==3.6.3
mistune==2.0.5
mlxtend==0.22.0
monai==1.1.0
multidict==6.0.4
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
nibabel==5.0.0
notebook-shim==0.2.2
notebook==6.5.2
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python-headless==4.7.0.68
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
panel==0.14.3
param==1.12.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.4.0
pip==23.1.2
pkgutil-resolve-name==1.3.10
platformdirs==3.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pyct==0.5.0
pydeprecate==0.3.2
pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
pytorch-lightning-spells==0.1.0
pytorch-lightning==1.6.5
pytz==2022.7.1
pyviz-comms==2.2.1
pywavelets==1.4.1
pyyaml==6.0
pyzmq==25.0.0
qudida==0.0.4
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rising==0.3.0
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
semantic-version==2.10.0
send2trash==1.8.0
sentry-sdk==1.13.0
setuptools==57.4.0
simpleitk==2.2.1
simplejson==3.18.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.11.2
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.2.1
torch==1.13.1
torchmetrics==0.7.3
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typing-extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.14
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.3.3
werkzeug==2.2.2
wheel==0.37.0
wrapt==1.14.1
wurlitzer==3.0.3
yarl==1.8.2
zipp==3.11.0

